changed restart App on DashBoardActivity, changed back to AsyncTask  for DB

+++ SqlCore/src/main/java/com/ravtech/david/sqlcore/DatabaseHelper.java
@@ -130,18 +130,21 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     public void onCreate(SQLiteDatabase db) {
 
         // creating required tables
-        db.execSQL(CREATE_TABLE_EVENTS);
+        if (db.isOpen())
+            db.execSQL(CREATE_TABLE_EVENTS);
 
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
 
-        // on upgrade drop older tables
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_EVENT);
+        if (db.isOpen()) {
+            // on upgrade drop older tables
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_EVENT);
 
-        // create new tables
-        onCreate(db);
+            // create new tables
+            onCreate(db);
+        }
     }
 
 
@@ -198,23 +201,31 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
         SQLiteDatabase db = this.getReadableDatabase();
 
-        Cursor c = db.rawQuery(selectQuery, null);
+        if (db.isOpen()) {
+            Cursor c = db.rawQuery(selectQuery, null);
 
-        // looping through all rows and adding to list
-        if (c.moveToFirst()) {
-            do {
-                Event event = convertRawToEvent(c);
+            // looping through all rows and adding to list
+            if (c.moveToFirst()) {
+                do {
+                    Event event = convertRawToEvent(c);
 
-                events.add(event);
-            } while (c.moveToNext());
+                    events.add(event);
+                } while (c.moveToNext());
+            }
+            c.close();
+            return events;
+        }else {
+            return new ArrayList<>();
         }
-
-        return events;
     }
 
     public ArrayList<Event> getListFromCursor(Cursor cursor) {
 
         ArrayList<Event> events = new ArrayList<>();
+
+        if (cursor == null){
+            return new ArrayList<>();
+        }
 
         if (cursor.moveToFirst()) {
             do {
@@ -236,11 +247,14 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
         SQLiteDatabase db = this.getReadableDatabase();
 
-        Cursor cursor = db.rawQuery(countQuery, null);
+        int count = 0;
+        if (db.isOpen()) {
+            Cursor cursor = db.rawQuery(countQuery, null);
 
-        int count = cursor.getCount();
+            count = cursor.getCount();
 
-        cursor.close();
+            cursor.close();
+        }
 
         // return count
         return count;
@@ -253,10 +267,12 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         String countQuery = "SELECT  * FROM " + TABLE_EVENT + " ORDER BY " + KEY_EVENT_ID + " DESC";
 
         SQLiteDatabase db = this.getReadableDatabase();
-
-        Cursor c = db.rawQuery(countQuery, null);
-
-        return c;
+        if (db.isOpen()) {
+            Cursor c = db.rawQuery(countQuery, null);
+            return c;
+        }else {
+            return null;
+        }
 
     }
 
@@ -294,9 +310,10 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                 " ORDER BY " + KEY_EVENT_ID + " DESC";
 
         SQLiteDatabase db = this.getReadableDatabase();
-
-        return db.rawQuery(countQuery, null);
-
+        if (db.isOpen()) {
+            return db.rawQuery(countQuery, null);
+        }
+        return null;
     }
     public Cursor getStopTypeShiftOrderByTimeFilterByDurationWithoutWork(int minEventDuration) {
         String countQuery = "SELECT  * FROM " + TABLE_EVENT +
@@ -306,9 +323,10 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                 " ORDER BY " + KEY_EVENT_ID + " DESC";
 
         SQLiteDatabase db = this.getReadableDatabase();
-
-        return db.rawQuery(countQuery, null);
-
+        if (db.isOpen()) {
+            return db.rawQuery(countQuery, null);
+        }
+        return null;
     }
 
     public Cursor getStopTypeShiftOrderByTimeFilterByDuration(int minEventDuration) {
